one-shot learning, readme, inclusive vocabular
most of the one-shot learning code is in place, updated readme, implemented code so that all codons and amino acids will be used in vocab even if they're not present in the training data. this rquires retraining all models, unfortunately. i also need to modify the cds extraction script so that it ensures that no ambiguity codes are left in the sequences either through removal or through codon imputation by referring to the protein sequence.